+++ packages/seahorse/src/components/ui/RatingStars.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import { View, Pressable } from "react-native-css/components";
+import { View } from "react-native-css/components";
+import { Pressable } from "../../primitives/pressable";
 import { Star } from "lucide-react-native";
 
 interface RatingStarsProps {